% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CimzMLParse}
\alias{CimzMLParse}
\title{imzML resulting data structure:}
\usage{
CimzMLParse(xml_path)
}
\arguments{
\item{xml_path:}{full path to imzML file}
}
\value{
UUID: a String object with the UUID
continuous_mode: a boolean which is true if spectra in continuous mode
compression_mz: a boolean indicating wheher data is compressed or not
compression_int: a boolean indicating wheher data is compressed or not
mz_dataType: a String with the data type: "float", "int"... etc...
int_dataType: a String with the data type: "float", "int"... etc...
pixel_size_um: a double with the pixel area? check this...
run_data: a data.frane with the columns: x, y, mzLength, mzOffset, intLength, intOffset
}
\description{
imzML resulting data structure:
}
