% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cload_imzMLImages}
\alias{Cload_imzMLImages}
\title{Load images}
\usage{
Cload_imzMLImages(
  rMSIobj,
  pixelIDs,
  massIndex,
  commonMassAxis,
  number_of_threads
)
}
\arguments{
\item{rMSIobj:}{an rMSI object prefilled with a parsed imzML.}

\item{pixelIDs:}{pixel ID's of the spectra to load in C-style indexing (starting at 0).}

\item{maxIndex:}{mass index range to load}

\item{commonMassAxis:}{a common mass axis that may be different than the mass axis in the rMSI object.}

\item{number_of_threads:}{number of thread to use during interpolation}
}
\value{
a matrix
}
\description{
Load images
}
