% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPeakMatrix.R
\name{getPeakMatrix}
\alias{getPeakMatrix}
\title{obtains the peak matrix from imzML files.}
\usage{
getPeakMatrix(
  data_file,
  params,
  lowMass = 0,
  highMass = 0,
  nThreads = 0,
  imzMLChecksum = F,
  imzMLSubCoords = NULL,
  fixBrokenUUID = F
)
}
\arguments{
\item{params}{"massResolution": mass resolution with which the spectra were acquired.
"SNR": signal-to-noise ratio
"noiseMethod": method for estimating noise.
"minPixelsSupport": minimum percentage of pixels that must support an ion for it to be considered.}

\item{data_file:}{absolute reference to the file with the imzML extension.}

\item{lowMass:}{lower mass to consider}

\item{highMass:}{higher mass to consider}

\item{nThreads:}{number of threads for parallel processing (if zero, nThreads=maxCores-1)}

\item{imzMLChecksum:}{if the binary file checksum must be verified, it can be disabled for convenice with really big files.}

\item{imzMLSubCoords:}{a Complex vector with the motors coordinates to be included in the ramdisk, if NULL all positions will be used.}

\item{fixBrokenUUID:}{set to FALSE by default to automatically fix an uuid mismatch between the ibd and the imzML files (a warning message will be raised).}
}
\value{
a list:
peakMatrix: Matrix of peak (centroids) rows = pixels, columns = intensity of each pixel.
mass: Vector with the masses associated with each column of peakMatrix.
pixelsSupport: Vector with the number of pixels in each column with non-zero intensity.
}
\description{
obtains the peak matrix from imzML files.
}
