% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imzMLBinReadAllMz}
\alias{imzMLBinReadAllMz}
\title{Generic method for the imzMLreader}
\usage{
imzMLBinReadAllMz(ibdFname, imzML)
}
\arguments{
\item{ibdFname:}{full path to the ibd file.}

\item{NPixels:}{Total number of pixels in the image.}

\item{N:}{number of elemetns (or data point to read).}

\item{offset:}{offset in bytes at which the reading operation is started.}

\item{read_mz:}{if true m/z data is readed, otherwise intensities are readed.}

\item{continuous:}{true if imzML data is in continuous mode}
}
\description{
Generic method for the imzMLreader
}
